add: weight model skeleton
