+++ lib/models/length/yard.dart
@@ -1,62 +1,62 @@
-import 'package:flutter_unit_converter/models/centimeter.dart';
-import 'package:flutter_unit_converter/models/feet.dart';
-import 'package:flutter_unit_converter/models/inch.dart';
-import 'package:flutter_unit_converter/models/kilometer.dart';
-import 'package:flutter_unit_converter/models/length_model.dart';
-import 'package:flutter_unit_converter/models/meter.dart';
-import 'package:flutter_unit_converter/models/micrometer.dart';
-import 'package:flutter_unit_converter/models/millimeter.dart';
-import 'package:flutter_unit_converter/models/nanometer.dart';
-import 'package:flutter_unit_converter/models/yard.dart';
+import 'package:flutter_unit_converter/models/length/centimeter.dart';
+import 'package:flutter_unit_converter/models/length/feet.dart';
+import 'package:flutter_unit_converter/models/length/inch.dart';
+import 'package:flutter_unit_converter/models/length/kilometer.dart';
+import 'package:flutter_unit_converter/models/length/length_model.dart';
+import 'package:flutter_unit_converter/models/length/meter.dart';
+import 'package:flutter_unit_converter/models/length/micrometer.dart';
+import 'package:flutter_unit_converter/models/length/mile.dart';
+import 'package:flutter_unit_converter/models/length/millimeter.dart';
+import 'package:flutter_unit_converter/models/length/nanometer.dart';
 
 /// class Centimeter with methods to convert from other length classes
-class Mile extends Length {
+class Yard extends Length {
   /// Constructor of the class, asigns the value
-  Mile(super.value);
+  Yard(super.value);
 
   /// returns an Inch object from another type object and do the unit convertion
   @override
-  factory Mile.from(Length len) {
+  factory Yard.from(Length len) {
     double val = len.value;
     switch (len.type) {
       case Nanometer:
-        val = val / (1609344 * 1000 * 1000);
+        val = val / (9144 * 100 * 1000);
         break;
       case Micrometer:
-        val = val / (1609344 * 1000);
+        val = val / (9144 * 100);
         break;
       case Millimeter:
-        val = val / 1609344;
+        val = val / (9144 / 10);
         break;
       case Centimeter:
-        val = val / (1609344 / 10);
+        val = val / (9144 / 10 / 10);
         break;
       case Inch:
-        val = val / 63360;
+        val = val / (3 * 12);
         break;
       case Feet:
-        val = val / 5280;
+        val = val / 3;
         break;
-      case Yard:
-        val = val / 1760;
+      case Mile:
+        val = val * 1760;
         break;
       case Meter:
-        val = val / (1609344 / 10 / 100);
+        val = val / (9144 / 100 / 100);
         break;
       case Kilometer:
-        val = val / (1609344 / 1000 / 1000);
+        val = val / (9144 / 100 / 100 / 1000);
     }
-    return Mile(val);
+    return Yard(val);
   }
 
   /// return the abreviation of the unit to be shown in the selectors
   @override
-  String get unit => 'mi';
+  String get unit => 'yd';
 
   /// returns the type of the object, we use it in the converter logic as
   /// we declare Length objects that can be of different types
   @override
-  Type get type => Mile;
+  Type get type => Yard;
 
   /// convert from inch object to the specified object
   @override
@@ -74,8 +74,8 @@ class Mile extends Length {
         return Inch.from(this);
       case Feet:
         return Feet.from(this);
-      case Yard:
-        return Yard.from(this);
+      case Mile:
+        return Mile.from(this);
       case Meter:
         return Meter.from(this);
       case Kilometer:
