+++ lib/models/weight/weight_model.dart
@@ -1,0 +1,11 @@
+abstract class Weight {
+  double value;
+  Weight(this.value);
+  // factory Weight.from(Type type, double value) {
+  //   return
+  // }
+
+  String get unit => '';
+  Type get type => Weight;
+  Weight to(Type type) => this;
+}
